# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For OrderSpack PHP Test",
        "description": "This is a demonstration of OrderSpack Php test.",
        "contact": {
            "name": "Developer",
            "url": "https://boosurostephen.com",
            "email": "boosurostephen@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "PHP TEST API Server"
        }
    ],
    "paths": {
...
        "/api/v1/variantgroups": {
            "get": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Fetch All Variant Groups",
                "description": "Get list of Variant Groups",
                "operationId": "getAllVariantGroups",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Store a variant group",
                "description": "Store Variant Group",
                "operationId": "storeVariantGroup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "variant_group_name": {
                                        "description": "Variant Group Name",
                                        "type": "string",
                                        "example": "Weight"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string",
                                        "example": "lorem ipsum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "variant group created successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/variantgroups/{id}": {
            "get": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Fetch a Variant Group",
                "description": "Fetch a variant group",
                "operationId": "getVariantGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Variant Group id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "variant group retrieved successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Update a variant group",
                "description": "Update Variant Group",
                "operationId": "updateVariantGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Variant Group id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "variant_group_name",
                        "in": "query",
                        "description": "Variant Group Name",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Color"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "schema": {
                            "type": "string"
                        },
                        "example": "lorem ipsum"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "variant group updated successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Variant Groups"
                ],
                "summary": "Delete Variant Group",
                "description": "Delete a variant group",
                "operationId": "deleteVariantGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Variant Group id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Variant Group deleted successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}